+++ MusicMover/Rules/MoveNewFileRule.cs
@@ -51,7 +51,10 @@ public class MoveNewFileRule : Rule
             !string.Equals(StateObject.MediaHandler.FileInfo.FullName, newFromFilePath))
         {
             MediaFileHelper.DumpCoverArt(StateObject.MediaHandler, StateObject.ToAlbumDirInfo, StateObject.Options);
-            StateObject.MediaHandler.FileInfo.Delete();
+            if (!StateObject.Options.KeepSourceFile)
+            {
+                StateObject.MediaHandler.FileInfo.Delete();
+            }
         }
 
         if (saveSuccess)
